requirements.txt
# Memory Layer Backend Dependencies
# Built with Kiro for Kiroween 2025 🎃

# FastAPI and server
fastapi==0.115.0
uvicorn==0.31.0
python-multipart==0.0.12
python-dotenv==1.2.1

# Pydantic for data validation
pydantic==2.12.4

# OpenAI for embeddings and LLM
openai==2.6.1

# FAISS for vector search
faiss-cpu==1.12.0
numpy==1.26.4

# Supabase for database and auth
supabase==2.24.0
PyJWT==2.10.1

# Stripe for payments (optional)
stripe==14.0.0

# AWS S3 for storage (optional)
boto3==1.40.61

# HTTP client
httpx==0.28.1
requests==2.32.5
